+++ core/scheduler.strategy.factory.simplestratfactory/src/de/uniol/inf/is/odysseus/scheduler/strategy/factory/simplestratfactory/BiggestQueueFactory.java
@@ -18,7 +18,7 @@ public class BiggestQueueFactory extends AbstractSchedulingFactory {
 	}
 
 	@Override
-	public IScheduling create(IPartialPlan plan, int priority) {
+	public IScheduling create(IPartialPlan plan, long priority) {
 		return new BiggestQueueScheduling(plan);
 	}
