+++ scheduling/scheduler.strategy.factory.chainfactory/src/de/uniol/inf/is/odysseus/scheduler/strategy/factory/chainfactory/ChainSchedulingFactory.java
@@ -16,7 +16,7 @@ public class ChainSchedulingFactory extends AbstractSchedulingFactory {
 	}
 
 	@Override
-	public IScheduling create(IPartialPlan plan, int priority) {
+	public IScheduling create(IPartialPlan plan, long priority) {
 		return new ChainScheduling(plan);
 	}
